applications/graph/GNN/PyTorch_Implementation/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
ase=3.21.1=pypi_0
blas=1.0=mkl
boost=1.74.0=py37h6dcda5c_3
boost-cpp=1.74.0=h9359b55_0
bzip2=1.0.8=h7f98852_4
ca-certificates=2020.12.5=ha878542_0
cached-property=1.5.2=pypi_0
cairo=1.16.0=h9f066cc_1006
certifi=2020.12.5=py37h89c1867_1
chardet=4.0.0=pypi_0
cudatoolkit=10.2.89=h8f6ccaa_8
cycler=0.10.0=py_2
decorator=4.4.2=pypi_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=hba837de_1004
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=h0b5b191_1005
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
googledrivedownloader=0.4=pypi_0
greenlet=1.0.0=py37hcd2ae1e_0
h5py=3.2.1=pypi_0
icu=67.1=he1b5a44_0
idna=2.10=pypi_0
importlib-metadata=3.10.0=py37h89c1867_0
isodate=0.6.0=pypi_0
jinja2=2.11.3=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.1=py37h2527ec5_1
lame=3.100=h7f98852_1001
ld_impl_linux-64=2.35.1=hea4e1c9_2
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libglib=2.68.0=h3e27bee_2
libiconv=1.16=h516909a_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.0.9=he6b73bb_1
libuuid=2.32.1=h7f98852_1000
libuv=1.41.0=h7f98852_0
libxcb=1.13=h7f98852_1003
libxml2=2.9.10=h68273f3_2
littleutils=0.2.2=pypi_0
llvm-openmp=11.1.0=h4bd325d_0
llvmlite=0.36.0=pypi_0
lz4-c=1.9.3=h9c3ff4c_0
markupsafe=1.1.1=pypi_0
matplotlib=3.4.1=pypi_0
matplotlib-base=3.2.2=py37h1d35a4c_1
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py37h8f50634_2
mkl_fft=1.3.0=py37h902c9e0_1
mkl_random=1.2.0=py37h9fdb41a_1
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
networkx=2.5=pypi_0
ninja=1.10.2=h4bd325d_0
numba=0.53.1=pypi_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
ogb=1.3.0=pypi_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openssl=1.1.1k=h7f98852_0
outdated=0.2.1=pypi_0
pandas=1.2.3=pypi_0
pcre=8.44=he1b5a44_0
pillow=8.2.0=pypi_0
pip=21.0.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
pthread-stubs=0.4=h36c2ea0_1001
pycairo=1.20.0=py37h01af8b0_1
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.7.10=hffdb5ce_100_cpython
python-dateutil=2.8.1=py_0
python-louvain=0.15=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.8.0=py3.7_cuda10.2_cudnn7.6.5_0
pytz=2021.1=pyhd8ed1ab_0
rdflib=5.0.0=pypi_0
rdkit=2021.03.1=py37haf5a968_0
readline=8.0=he28a2e2_2
reportlab=3.5.66=py37h69800bb_0
requests=2.25.1=pypi_0
scikit-learn=0.24.1=pypi_0
scipy=1.6.2=pypi_0
setuptools=49.6.0=py37h89c1867_3
six=1.15.0=pyh9f0ad1d_0
sqlalchemy=1.4.4=py37h5e8e339_0
sqlite=3.35.3=h74cdb3f_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=h21135ba_1
torch-cluster=1.5.9=pypi_0
torch-geometric=1.6.3=pypi_0
torch-scatter=2.0.6=pypi_0
torch-sparse=0.6.9=pypi_0
torch-spline-conv=1.2.1=pypi_0
torchaudio=0.8.0=py37
torchvision=0.9.0=py37_cu102
tornado=6.1=py37h5e8e339_1
tqdm=4.59.0=pypi_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.4=pypi_0
wheel=0.36.2=pyhd3deb0d_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.0=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h516909a_1010
zstd=1.4.9=ha95c52a_0
